Fix warnings from splint.
As part of the finger exercises to get into a better programmer, I am
using various lint tools. Splint showed that I was comparing different
types and also not looking at the return from certain functions. Since
that could lead to issues depending on the system, I was more explicit
in what was wanted.

+++ k_r/chap_1/ex04_c2f.c
@@ -16,7 +16,7 @@ int main(){
 
     celsius = (double) lower;
     printf("Celsius to Fahrenheit Conversion\n");
-    while (celsius <= upper) {
+    while (celsius <= (double) upper) {
 	fahr = (( 9.0 / 5.0 ) * celsius ) +32.0;
 	printf("%3.0f\t=>\t%6.1f\n",celsius,fahr);
 	celsius = celsius + (double) step;
